feat: 音频播放器字幕动画提升；番剧分级列表查看；系统消息查看；全局浮动字幕；实验性功能；评论详情查看 chore: 部分设置项ui；部分图标 fix: 一大堆问题（忘了

+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/bangumi/info/remote/StatXX.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.bilibilisdk.sdk.bangumi.info.remote
 
 data class StatXX(
-    val coins: Int,
+    val coins: Double,
     val danmakus: Int,
     val favorite: Int,
     val favorites: Int,
